==> 02-YouTubeNotebooks/01-SelectColumns.ipynb <==
# IMPORTS

## Libraries

import re
import inflection

import pandas as pd
import numpy  as np


pd.set_option('display.max_columns', 200)

## Load Data

dfRaw1 = pd.read_csv('../Data/youtube_dlc/YouTubeStackExchangeQueryRaw.csv', parse_dates=['upload_date'])

dfRaw1.sample(5)

# DESCRIPTION OF THE DATA

dfRaw2 = dfRaw1.copy()

dfRaw2.columns

## Rename Columns

camelize = lambda col: inflection.camelize(col)

newColumns = list(map(camelize, dfRaw2.columns))

# rename
dfRaw2.columns = newColumns

dfRaw2.columns

## Columns Filter

columns = ['Id', 'Uploader', 'UploadDate', 'Title', 'Thumbnails', 'Description', 'Categories', 'Tags',
       'Duration', 'ViewCount', 'LikeCount', 'DislikeCount', 'AverageRating', 'Query', 'DaysSincePublication']

dfRaw2 = dfRaw2[columns]

## Extracting information from the original columns

#Categories
dfRaw2['Categories'] = dfRaw2['Categories'].apply(lambda row: row.strip("'[").strip("']"))
#Tags
dfRaw2['Tags'] = dfRaw2['Tags'].apply(lambda row: str(row).strip("[").strip("]").replace("'",''))

#Thumbanil URL
def GetThumbnail(text):
    pattern = "http[s]?://(?:[a-zA-Z]|[0-9]|[$-_@.&+]|[!*\(\),]|(?:%[0-9a-fA-F][0-9a-fA-F]))+"
    urlThumbnail = re.search(pattern, text).group(0).rstrip("'")
    return urlThumbnail

dfRaw2['ThumbnailUrl'] = dfRaw2['Thumbnails'].apply(lambda row: GetThumbnail(str(row).strip('"[').strip(']"').split(',')[0]))

# Thumbnail Width and Height
def GetValue(text):
    pattern = "(\d+)"
    value = int(re.search(pattern, text).group(0))
    return value

# Thumbnail Width
dfRaw2['ThumbnailWidth'] = dfRaw2['Thumbnails'].apply(lambda row: GetValue(str(row).strip('"[').strip(']"').split(',')[1]))
# Thumbnail Height
dfRaw2['ThumbnailHeight'] = dfRaw2['Thumbnails'].apply(lambda row: GetValue(str(row).strip('"[').strip(']"').split(',')[2]))

### Drop columns irrelevant

dfRaw2 = dfRaw2.drop('Thumbnails', axis=1)

# Convert to .feather

dfRaw2.to_feather('../Data/FeatherData/YouTubeStackExchangeQueryRaw.feather')
