==> 02-YouTubeNotebooks/02-DescriptionDataAndGoogleSheetsSend.ipynb <==
# IMPORTS

## Libraries

import math

import pandas as pd
import numpy  as np

from google.oauth2             import service_account
from googleapiclient.discovery import build



pd.set_option('display.max_columns', 200)

## Load Data

dfRaw1 = pd.read_feather('../Data/FeatherData/YouTubeStackExchangeQueryRaw.feather')

dfRaw1.sample(5)

# DESCRIPTION OF THE DATA

print(f'Number of Rows: {dfRaw1.shape[0]}')
print(f'Number of Columns: {dfRaw1.shape[1]}')

## Fillout NA

dfRaw1.isnull().sum()

# LikeCount
dfRaw1['LikeCount'] = dfRaw1['LikeCount'].apply(lambda row: 0 if math.isnan(row) else row)

# DislikeCount
dfRaw1['DislikeCount'] = dfRaw1['DislikeCount'].apply(lambda row: 0 if math.isnan(row) else row)

# DislikeCount
dfRaw1['AverageRating'] = dfRaw1['AverageRating'].apply(lambda row: -1 if math.isnan(row) else row)

# Description
dfRaw1['Description'].fillna("no description", inplace=True)

dfRaw1.isnull().sum()

# Shuffle Data and Prepare to send to Google Sheets

dfShuffle = dfRaw1.sample(frac=1)
dfShuffle['UploadDate'] = dfShuffle['UploadDate'].astype(str)
dfShuffle = dfShuffle.values.tolist()

# Send to Google Sheets

# Documentation: https://developers.google.com/sheets/api/quickstart/python
SERVICE_ACCOUNT_FILE = 'D:/01-DataScience/04-Projetos/00-Git/Youtube-Video-Recommendations/Credentials/keys.json'
SCOPES = ['https://www.googleapis.com/auth/spreadsheets']

credentials = None
credentials = service_account.Credentials.from_service_account_file(
        SERVICE_ACCOUNT_FILE, scopes=SCOPES)


# The ID of spreadsheet.
SAMPLE_SPREADSHEET_ID = '1uCur7jOXuLnwuwfWgoBL8mvDDvchuLf-o0X-AnOxS7s'

service = build('sheets', 'v4', credentials=credentials)


# Call the Sheets API
sheet = service.spreadsheets()

# Read Values
# result = sheet.values().get(spreadsheetId=SAMPLE_SPREADSHEET_ID,
#                             range="YoutubeScrapeSample!A1:B11").execute()

# values = result.get('values', [])

# Write Values
request = sheet.values().update(spreadsheetId=SAMPLE_SPREADSHEET_ID,
                               range="YoutubeScrapeSample!A2", valueInputOption="USER_ENTERED", body={"values":dfShuffle}).execute()

dfRaw1.columns
