==> 02-YouTubeNotebooks/03-GoogleSheetsRead.ipynb <==
# IMPORTS

## Libraries

import math

import pandas as pd
import numpy  as np

from google.oauth2             import service_account
from googleapiclient.discovery import build


pd.set_option('display.max_columns', 200)

# Read Google Sheets

# Documentation: https://developers.google.com/sheets/api/quickstart/python
SERVICE_ACCOUNT_FILE = 'D:/01-DataScience/04-Projetos/00-Git/Youtube-Video-Recommendations/Credentials/keys.json'
SCOPES = ['https://www.googleapis.com/auth/spreadsheets']

credentials = None
credentials = service_account.Credentials.from_service_account_file(
        SERVICE_ACCOUNT_FILE, scopes=SCOPES)


# The ID of spreadsheet.
SAMPLE_SPREADSHEET_ID = '1uCur7jOXuLnwuwfWgoBL8mvDDvchuLf-o0X-AnOxS7s'

service = build('sheets', 'v4', credentials=credentials)


# Call the Sheets API
sheet = service.spreadsheets()

#Read Values
result = sheet.values().get(spreadsheetId=SAMPLE_SPREADSHEET_ID,
                            range="YoutubeScrapeSample!A1:R").execute()

values = result.get('values', [])

dfGoogleSheets = pd.DataFrame(values[1:], columns=values[0])

dfGoogleSheets['UploadDate'] = pd.to_datetime(dfGoogleSheets['UploadDate'])
dfGoogleSheets['WatchList'] = dfGoogleSheets['WatchList'].replace('', np.nan)
dfGoogleSheets[['DaysSincePublication', 'Duration', 'ViewCount', 'LikeCount', 'DislikeCount']] = dfGoogleSheets[['DaysSincePublication', 'Duration', 'ViewCount', 'LikeCount', 'DislikeCount']].astype(int)
dfGoogleSheets['AverageRating'] = dfGoogleSheets['AverageRating'].astype(float)
dfGoogleSheets[dfGoogleSheets.select_dtypes(include=['object']).columns] = dfGoogleSheets.select_dtypes(include=['object']).astype('category')

dfGoogleSheets.dtypes

## Drop Duplicates

dfGoogleSheets = dfGoogleSheets.drop_duplicates(subset=['Id'])

dfGoogleSheets.shape

## Split into Labeled and Unlabeled Dataset

dfTrain = dfGoogleSheets[dfGoogleSheets['WatchList'].notnull()].reset_index(drop=True)
dfTest = dfGoogleSheets[dfGoogleSheets['WatchList'].isnull()].reset_index(drop=True)

# Train Dataset
print('\n\tTrain Dataset')
print(f'Number of Rows: {dfTrain.shape[0]}')
print(f'Number of Columns: {dfTrain.shape[1]}')

# Test Dataset
print('\n\n\tTest Dataset')
print(f'Number of Rows: {dfTest.shape[0]}')
print(f'Number of Columns: {dfTest.shape[1]}')

# Convert to .feather

dfTrain.to_feather('../Data/FeatherData/dfTrainGS.feather')
dfTest.to_feather('../Data/FeatherData/dfTestGS.feather')
