==> 02-YouTubeNotebooks/06-ActiveLearningSecondPart.ipynb <==
# IMPORTS

## Libraries

import pandas as pd
import numpy  as np

import matplotlib.pyplot as plt

from scipy.sparse                    import hstack
from sklearn.metrics                 import roc_auc_score, average_precision_score
from sklearn.ensemble                import RandomForestClassifier
from sklearn.feature_extraction.text import TfidfVectorizer


np.random.seed(0)
%matplotlib inline
pd.set_option('display.max_columns', 200)

# Load Data

dfRaw1 = pd.read_feather('../Data/FeatherData/dfTrainGS.feather')
dfRaw2 = pd.read_feather('../Data/FeatherData/dfActiveLearningGS.feather')

dfRaw1 = dfRaw1.sort_values('UploadDate')
dfRaw2 = dfRaw2.sort_values('UploadDate')

dfRaw1['WatchList'] = dfRaw1['WatchList'].astype(int)
dfRaw2['WatchList'] = dfRaw2['WatchList'].astype(int)

# Active Learning Results

dfRaw2['New'] = 1

dfRaw2.shape

average_precision_score(dfRaw2['WatchList'], dfRaw2['p']), roc_auc_score(dfRaw2['WatchList'], dfRaw2['p'])

dfTrain = pd.concat([dfRaw1, dfRaw2.drop('p', axis=1)])

dfTrain['New'] = dfTrain['New'].fillna(0)

dfFeatures = pd.DataFrame(index=dfTrain.index)
dfFeatures['ViewCount'] = dfTrain['ViewCount']
dfFeatures['DaysSincePublication'] = dfTrain['DaysSincePublication']
dfFeatures['WatchList'] = dfTrain['WatchList']
dfFeatures['ViewsPerDay'] = dfFeatures['ViewCount'] / dfFeatures['DaysSincePublication']
dfFeatures = dfFeatures.drop('DaysSincePublication', axis=1)

# Split DataFrame into Training and Validation Dataset

maskTrain = (dfTrain['UploadDate'] < '2019-10-23')

maskVal = (dfTrain['UploadDate'] >= '2019-10-23') & (dfTrain['New'] == 0)

Xtrain, Xval = dfFeatures[maskTrain].drop('WatchList', axis=1) , dfFeatures[maskVal].drop('WatchList', axis=1)
ytrain, yval = dfFeatures[maskTrain]['WatchList'], dfFeatures[maskVal]['WatchList']

Xtrain.shape, Xval.shape, ytrain.shape, yval.shape

# Text Features

titleTrain = dfTrain[maskTrain]['Title']
titleVal = dfTrain[maskVal]['Title']

titleVec = TfidfVectorizer(min_df=2)
titleBowTrain = titleVec.fit_transform(titleTrain)
titleBowVal = titleVec.transform(titleVal)

titleBowTrain.shape

titleBowTrain

XtrainWTitle = hstack([Xtrain, titleBowTrain])
XvalWTitle = hstack([Xval, titleBowVal])

XtrainWTitle.shape, XvalWTitle.shape

# Model

## RandomForestClassifier

model = RandomForestClassifier(n_estimators=1000, random_state=0, class_weight='balanced', n_jobs=-1)
model.fit(XtrainWTitle, ytrain)

p = model.predict_proba(XvalWTitle)[:,1]

## Model Evaluate 

average_precision_score(yval,p)

roc_auc_score(yval, p)

# Convert Dataset to .feather

dfTrain = dfTrain.reset_index(drop=True)

dfTrain.to_feather('../Data/FeatherData/dfAllDataLabeled.feather')
