==> 02-YouTubeNotebooks/07-Modelling.ipynb <==
# IMPORTS

## Libraries

import warnings

import pandas as pd
import numpy  as np

import matplotlib.pyplot as plt

from skopt                           import forest_minimize
from lightgbm                        import LGBMClassifier
from scipy.sparse                    import hstack, csr_matrix
from sklearn.metrics                 import roc_auc_score, average_precision_score
from sklearn.ensemble                import RandomForestClassifier
from sklearn.linear_model            import LogisticRegression
from sklearn.preprocessing           import MaxAbsScaler
from sklearn.feature_extraction.text import TfidfVectorizer


np.random.seed(0)
%matplotlib inline
pd.set_option('display.max_columns', 200)
warnings.filterwarnings("ignore")

# Load Data

dfRaw = pd.read_feather('../Data/FeatherData/dfAllDataLabeled.feather')

dfRaw = dfRaw.sort_values('UploadDate')

# Feature Engineering

dfRaw.shape

dfFeatures = pd.DataFrame(index=dfRaw.index)
dfFeatures['ViewCount'] = dfRaw['ViewCount']
dfFeatures['DaysSincePublication'] = dfRaw['DaysSincePublication']
dfFeatures['WatchList'] = dfRaw['WatchList']
dfFeatures['ViewsPerDay'] = dfFeatures['ViewCount'] / dfFeatures['DaysSincePublication']
dfFeatures = dfFeatures.drop('DaysSincePublication', axis=1)

# Split DataFrame into Training and Validation Dataset

dfRaw['UploadDate'].iloc[int(round(dfRaw.shape[0]/2,0))]

dateSplit = '2019-11-27'

maskTrain = (dfRaw['UploadDate'] < dateSplit)

maskVal = (dfRaw['UploadDate'] >= dateSplit)

Xtrain, Xval = dfFeatures[maskTrain].drop('WatchList', axis=1) , dfFeatures[maskVal].drop('WatchList', axis=1)
ytrain, yval = dfFeatures[maskTrain]['WatchList'], dfFeatures[maskVal]['WatchList']

Xtrain.shape, Xval.shape, ytrain.shape, yval.shape

# Text Features

titleTrain = dfRaw[maskTrain]['Title']
titleVal = dfRaw[maskVal]['Title']

titleVec = TfidfVectorizer(min_df=4, ngram_range=(1,2))
titleBowTrain = titleVec.fit_transform(titleTrain)
titleBowVal = titleVec.transform(titleVal)

titleBowTrain.shape

titleBowTrain

XtrainWTitle = hstack([Xtrain, titleBowTrain])
XvalWTitle = hstack([Xval, titleBowVal])

XtrainWTitle.shape, XvalWTitle.shape

# Model

## RandomForestClassifier

rf = RandomForestClassifier(n_estimators=1500, random_state=0, min_samples_leaf=5, class_weight='balanced', n_jobs=-1)
rf.fit(XtrainWTitle, ytrain)

pRf = rf.predict_proba(XvalWTitle)[:,1]

### Model Evaluate 

print(f'ap: {average_precision_score(yval, pRf)}, auc: {roc_auc_score(yval, pRf)}')

# ap: 0.3521564476834772, auc: 0.7758089368258859 | min_samples_leaf=1
# ap: 0.3843011851670597, auc: 0.7815540391811577 | min_samples_leaf=2
# ap: 0.38551950447476385, auc: 0.7831168831168831 | n_estimators=1500
# ap: 0.40864867634952645, auc: 0.7861765353290777 | min_df=3
# ap: 0.4138684826920506, auc: 0.784668721109399 | ngram_range = (1,2)
# ap: 0.425697264360267, auc: 0.7845256438476778 | min_df=4 & ngram_range = (1,2)

### Bayesian Optimization

def tuneRf(params):
    print(params)
    max_depth = params[0]
    min_samples_leaf = params[1]
    n_estimators = params[2]
    min_df = params[3]
    ngram_range = (1, params[4])
    
    titleVec = TfidfVectorizer(min_df=min_df, ngram_range=ngram_range)
    titleBowTrain = titleVec.fit_transform(titleTrain)
    titleBowVal = titleVec.transform(titleVal)
    
    XtrainWTitle = hstack([Xtrain, titleBowTrain])
    XvalWTitle = hstack([Xval, titleBowVal])
    
    model = RandomForestClassifier(n_estimators=n_estimators, max_depth=max_depth, min_samples_leaf=min_samples_leaf, random_state=0,
                                   class_weight='balanced', n_jobs=-1)
    
    model.fit(XtrainWTitle, ytrain)
    
    p = model.predict_proba(XvalWTitle)[:,1]
    
    print(roc_auc_score(yval, p))
    
    return -average_precision_score(yval,p)



space = [(1, 10), #max_depth
        (1, 20), #min_samples_leaf 
        (100, 1000), #n_estimetors
        (1, 5), #min_df
        (1, 5)] #ngram_range



result = forest_minimize(tuneRf, space, random_state=160745, n_random_starts=20, n_calls=50, verbose=1)

result.x

### Model With Hyperparameter Tuning

params = [8, #max_depth
 19, #min_samples_leaf 
 840, #n_estimetors
 3,#min_df
 5] #ngram_range


max_depth = params[0]
min_samples_leaf = params[1]
n_estimators = params[2]
min_df = params[3]
ngram_range = (1, params[4])


titleVec = TfidfVectorizer(min_df=min_df, ngram_range=ngram_range)
titleBowTrain = titleVec.fit_transform(titleTrain)
titleBowVal = titleVec.transform(titleVal)

XtrainWTitle = hstack([Xtrain, titleBowTrain])
XvalWTitle = hstack([Xval, titleBowVal])

modelRf = RandomForestClassifier(n_estimators=n_estimators, max_depth=max_depth, min_samples_leaf=min_samples_leaf, random_state=0,
                               class_weight='balanced', n_jobs=-1)

modelRf.fit(XtrainWTitle, ytrain)

pRf = modelRf.predict_proba(XvalWTitle)[:,1]


print(f'ap: {average_precision_score(yval,pRf)}, auc: {roc_auc_score(yval, pRf)}')

## LGBMClassifier

modelLGBM = LGBMClassifier(random_state=0, class_weight='balanced', n_jobs=-1)
modelLGBM.fit(XtrainWTitle, ytrain)

pLGBM = modelLGBM.predict_proba(XvalWTitle)[:,1]

### Model Evaluate 

print(f'ap: {average_precision_score(yval,pLGBM)}, auc: {roc_auc_score(yval, pLGBM)}')

### Bayesian Optimization

def tuneLGBM(params):
    print(params)
    lr = params[0]
    max_depth = params[1]
    min_child_samples = params[2]
    subsample = params[3]
    colsample_bytree = params[4]
    n_estimators = params[5]
    min_df = params[6]
    ngram_range = (1, params[7])
    
    titleVec = TfidfVectorizer(min_df=min_df, ngram_range=ngram_range)
    titleBowTrain = titleVec.fit_transform(titleTrain)
    titleBowVal = titleVec.transform(titleVal)
    
    XtrainWTitle = hstack([Xtrain, titleBowTrain])
    XvalWTitle = hstack([Xval, titleBowVal])
    
    model = LGBMClassifier(learning_rate=lr, num_leaves=2 ** max_depth, max_depth=max_depth,
                           min_child_samples=min_child_samples, subsample=subsample, colsample_bytree=colsample_bytree,
                           bagging_freq=1, n_estimators=n_estimators, random_state=0, class_weight='balanced', n_jobs=-1)
    
    model.fit(XtrainWTitle, ytrain)
    
    p = model.predict_proba(XvalWTitle)[:,1]
    
    print(roc_auc_score(yval, p))
    
    return -average_precision_score(yval,p)



space = [(1e-3, 1e-1, 'log-uniform'), #lr
        (1, 10), #max_depth
        (1, 20), #min_child_samples
        (0.05, 1.0), #subsample
        (0.05, 1.0), #colsample_bytree
        (100, 1000), #n_estimetors
        (1, 5), #min_df
        (1, 5)] #ngram_range



result = forest_minimize(tuneLGBM, space, random_state=160745, n_random_starts=20, n_calls=50, verbose=1)

result.x

##### LGBMClassifier

- [0.001477438108754858, 3, 10, 0.16137221087324535, 0.19593296243176872, 966, 3, 5]
- ap: 0.4227, auc: 0.7806845696676206

## Model With Hyperparameter Tuning

params = [0.001477438108754858, #lr
 3, #max_depth
 10, #min_child_samples
 0.16137221087324535, #subsample
 0.19593296243176872, #colsample_bytree
 966, #n_estimetors
 3,  #min_df
 5] #ngram_range



lr = params[0]
max_depth = params[1]
min_child_samples = params[2]
subsample = params[3]
colsample_bytree = params[4]
n_estimators = params[5]
min_df = params[6]
ngram_range = (1, params[7])

titleVec = TfidfVectorizer(min_df=min_df, ngram_range=ngram_range)
titleBowTrain = titleVec.fit_transform(titleTrain)
titleBowVal = titleVec.transform(titleVal)

XtrainWTitle = hstack([Xtrain, titleBowTrain])
XvalWTitle = hstack([Xval, titleBowVal])

modelLgbm = LGBMClassifier(learning_rate=lr, num_leaves=2 ** max_depth, max_depth=max_depth,
                       min_child_samples=min_child_samples, subsample=subsample, colsample_bytree=colsample_bytree,
                       bagging_freq=1, n_estimators=n_estimators, random_state=0, class_weight='balanced', n_jobs=-1)

modelLgbm.fit(XtrainWTitle, ytrain)

pLgbm = modelLgbm.predict_proba(XvalWTitle)[:,1]


print(f'ap: {average_precision_score(yval,pLgbm)}, auc: {roc_auc_score(yval, pLgbm)}')

#ap: 0.4211858121788448, auc: 0.7803543913713407

# Logistic Regression

### Data Preparation

XtrainWTitle2 = csr_matrix(XtrainWTitle.copy())
XvalWTitle2 = csr_matrix(XvalWTitle.copy())


scaler = MaxAbsScaler()

XtrainWTitle2 = scaler.fit_transform(XtrainWTitle2)
XvalWTitle2 = scaler.transform(XvalWTitle2)

XvalWTitle2.shape

### Model

modelLr = LogisticRegression(C=0.01, n_jobs=-1, random_state=0)
modelLr.fit(XtrainWTitle2, ytrain)

pLr = modelLr.predict_proba(XvalWTitle2)[:,1]

print(f'ap: {average_precision_score(yval, pLr)}, auc: {roc_auc_score(yval, pLr)}')

# Ensamble

#### min_df > 3
- LGBM | ap: 0.42266615524449913, auc: 0.7806845696676206
- LR   | ap: 0.3986324868099709, auc: 0.7804974686330618
- RF   | ap: 0.43734417259145686, auc: 0.788895003301783

#### min_df > 2
- LGBM | ap: 0.41114784456628284, auc: 0.7786154523442659
- LR   | ap: 0.3809915480394665, auc: 0.7685009905348889
- RF   | ap: 0.4035174762881823, auc: 0.7769480519480519

pd.DataFrame({'RF': pRf,'LR': pLr, 'LGBM': pLgbm}).corr()

p = (0.56*pRf + 0.44*pLgbm)
print(f'ap: {average_precision_score(yval,p)}, auc: {roc_auc_score(yval, p)}')

# ap: 0.4325093482032139, auc: 0.7893132291437376
# ap: 0.4333159558286226, auc: 0.7887739379264804

# Save Model

import joblib as jb

# jb.dump(modelRf,'D:/01-DataScience/04-Projetos/00-Git/Youtube-Video-Recommendations/Model/modelRf.pkl.z')

# jb.dump(modelRf,'D:/01-DataScience/04-Projetos/00-Git/Youtube-Video-Recommendations/Model/modelRf.pkl.z')
# jb.dump(modelLgbm,'D:/01-DataScience/04-Projetos/00-Git/Youtube-Video-Recommendations/Model/modelLgbm.pkl.z')
# jb.dump(modelLr,'D:/01-DataScience/04-Projetos/00-Git/Youtube-Video-Recommendations/Model/modelLgbm.pkl.z')
